FinnhubProducer/requirements.txt
finnhub-python==2.4.20
websocket-client==1.8.0
confluent-kafka==2.5.3
avro-python3==1.10.2